repository

+++ src/main/java/org/example/sptingsecurityauthorizationaaa/controller/UserController.java
@@ -58,7 +58,7 @@ public class UserController {
 
     /* ---------------- DELETE USER ------------------------ */
     @RequestMapping(value = "/users/{id}", method = RequestMethod.DELETE)
-    public ResponseEntity<String> deleteUserById(@PathVariable int id) {
+    public ResponseEntity<String> deleteUserById(@PathVariable long id) {
         userService.delete(id);
         return new ResponseEntity<>("Deleted!", HttpStatus.OK);
     }
